app/requirements.txt
Django==4.0.2
mysqlclient==2.1.0
requests==2.27.1
celery==5.2.3
redis==4.1.2
django-debug-toolbar==3.2.4
django-celery-beat==2.1.0
django-celery-results==2.2.0
gunicorn==20.1.0
django-redis-cache==1.8.0
social-auth-app-django==5.0.0
django-crispy-forms
crispy-bootstrap5
django-hijack
python-dateutil